ProTecHub/pyproject.toml
[tool.poetry]
name = "ProTecHub"
version = "1.0.0"
description = "Projeto em Django sobre Gerenciamento de EPI's"
authors = ["Fabrício dos Santos Moreira", "Guilherme Stadnicki da Silva", "Maria Eduarda Figueiredo"]
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.10"
Django = "5.1.2"
pytz = "2024.2"
django-environ = "0.11.2"
mysqlclient = "2.2.5"
pillow = "11.0.0"
python-dateutil = "2.9.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
